Updated to new version + test fixe of include

+++ courses/cunix2/libft/src/ft_strncmp.c
@@ -2,7 +2,7 @@
 
 int ft_strncmp(const char *string1, const char *string2, size_t num)
 {
-    for (int i = 0; i < num; i++)
+    for (size_t i = 0; i < num; i++)
     {
         if (string1[i] < string2[i])
         {
